青と黒の輪郭線読み込みを使い分けられるようにした

+++ MainViewer_SerialSelector_SerialList.cpp
@@ -37,7 +37,7 @@ void MainViewer::SerialSelector::SerialList::update()
 	{
 		const auto items = getChildViewers<Item>();
 
-		getChildViewer<GUISlider>()->setHeight(items.size() * itemHeight);
+		getChildViewer<GUISlider>()->setHeight(double(items.size()) * itemHeight);
 
 		const auto delta = -getChildViewer<GUISlider>()->getDelta();
